Namespace APIs, Redesign OTP handling through events (#43)
