+++ src/orders/index.js
@@ -1,0 +1,7 @@
+import history from './history';
+import submit from './submit';
+
+export default {
+    ...history,
+    ...submit
+}
