+++ src/orders/submit.js
@@ -1,0 +1,163 @@
+import endpoints from '../api/endpoints';
+import { handleRequest } from '../network/https';
+
+export default {
+
+  /**
+   * Cancels the pending order specified by the order id.
+   *
+   * @param {*} orderId The pending order to cancel
+   */
+  cancel: async (orderId) => handleRequest(endpoints.CANCEL_ORDER, { orderId }),
+
+  /**
+   * Cancels all pending orders under the WealthSimple Trade Account.
+   *
+   * @param {*} accountId The specific account in the WealthSimple Trade account
+   */
+  cancelPending: async (accountId) => {
+    const pending = await wealthsimple.getPendingOrders(accountId);
+    return Promise.all(pending.orders.map(async (order) => await wealthsimple.cancelOrder(order.order_id)));
+  },
+
+  /**
+   * Market buy a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} quantity The number of securities to purchase
+   */
+  marketBuy: async (accountId, ticker, quantity) => {
+    let extensive_details = await wealthsimple.getSecurity(ticker, true);
+
+    return handleRequest(endpoints.PLACE_ORDER, {
+      security_id: extensive_details.id,
+      limit_price: extensive_details.quote.amount,
+      quantity,
+      order_type: "buy_quantity",
+      order_sub_type: "market",
+      time_in_force: "day",
+      account_id: accountId
+    });
+  },
+
+  /**
+   * Limit buy a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} limit The maximum price to purchase the security at
+   * @param {*} quantity The number of securities to purchase
+   */
+  limitBuy: async (accountId, ticker, limit, quantity) =>
+    handleRequest(endpoints.PLACE_ORDER, {
+      security_id: (await wealthsimple.getSecurity(ticker)).id,
+      limit_price: limit,
+      quantity,
+      order_type: "buy_quantity",
+      order_sub_type: "limit",
+      time_in_force: "day",
+      account_id: accountId
+    }),
+
+  /**
+   * Stop limit buy a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} stop The price at which the order converts to a limit order
+   * @param {*} limit The maximum price to purchase the security at
+   * @param {*} quantity The number of securities to purchase
+   */
+  stopLimitBuy: async (accountId, ticker, stop, limit, quantity) => {
+    
+    let security = await wealthsimple.getSecurity(ticker);
+
+    // The WealthSimple Trade backend doesn't check for this, even though the app does..
+    if (isCanadianSecurity(security.stock.primary_exchange) && stop !== limit) {
+      return Promise.reject({ reason: "TSX/TSX-V securities must have an equivalent stop and limit price." });
+    }
+
+    return handleRequest(endpoints.PLACE_ORDER, {
+      security_id: (await wealthsimple.getSecurity(ticker)).id,
+      stop_price: stop,
+      limit_price: limit,
+      quantity,
+      order_type: "buy_quantity",
+      order_sub_type: "stop_limit",
+      time_in_force: "day",
+      account_id: accountId
+    });
+  },
+
+  /**
+   * Market sell a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} quantity The number of securities to purchase
+   */
+  marketSell: async (accountId, ticker, quantity) => {
+
+    let extensive_details = await wealthsimple.getSecurity(ticker, true);
+    
+    return handleRequest(endpoints.PLACE_ORDER, {
+      security_id: extensive_details.id,
+      market_value: extensive_details.quote.amount,
+      quantity: quantity,
+      order_type: "sell_quantity",
+      order_sub_type: "market",
+      time_in_force: "day",
+      account_id: accountId,
+    });
+  },
+
+  /**
+   * Limit sell a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} limit The minimum price to sell the security at
+   * @param {*} quantity The number of securities to sell
+   */
+  limitSell: async (accountId, ticker, limit, quantity) =>
+    handleRequest(endpoints.PLACE_ORDER, {
+      security_id: (await wealthsimple.getSecurity(ticker)).id,
+      limit_price: limit,
+      quantity,
+      order_type: "sell_quantity",
+      order_sub_type: "limit",
+      time_in_force: "day",
+      account_id: accountId
+    }),
+
+  /**
+   * Stop limit sell a security through the WealthSimple Trade application.
+   *
+   * @param {*} accountId The account to make the transaction from
+   * @param {*} ticker The security symbol
+   * @param {*} stop The price at which the order converts to a limit order
+   * @param {*} limit The minimum price to sell the security at
+   * @param {*} quantity The number of securities to sell
+   */
+  stopLimitSell: async (accountId, ticker, stop, limit, quantity) => {
+
+    let security = await wealthsimple.getSecurity(ticker);
+
+    // The WealthSimple Trade backend doesn't check for this, even though the app does..
+    if (isCanadianSecurity(security.stock.primary_exchange) && stop !== limit) {
+      return Promise.reject({ reason: "TSX/TSX-V securities must have an equivalent stop and limit price." });
+    }
+
+    return handleRequest(endpoints.PLACE_ORDER, {
+      security_id: security.id,
+      stop_price: stop,
+      limit_price: limit,
+      quantity,
+      order_type: "sell_quantity",
+      order_sub_type: "stop_limit",
+      time_in_force: "day",
+      account_id: accountId
+    });
+  }
+};
